buy_item/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pyasn1==0.5.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
soupsieve==2.5
steampy==1.1.2
urllib3==2.1.0
yarl==1.9.4
